Fix a few warnings (unrefreenced variables, size_t conversions (migrating toward size_t)

+++ src/imglib/pngimage.h
@@ -11,9 +11,9 @@ IMAGE_NAMESPACE
 namespace loader {
 #endif
    // decompress a buffer into an image...
-	ImageFile *ImagePngFile (_8 * buf, long size);
+	ImageFile *ImagePngFile (_8 * buf, size_t size);
 	// compress image into a buffer
-	LOGICAL PngImageFile ( Image pImage, _8 ** buf, int *size);
+	LOGICAL PngImageFile ( Image pImage, _8 ** buf, size_t *size);
 #ifdef __cplusplus
 }
 #endif
